Update from master
This iteration has the following update:

Update the header and footer so that it is consistent with the design
provided by japan’s HQ. (Done)
Colour scheme as to be the same as the Japan HQ (Done)
Advised to have white background colour, but we might try to have the
faded sunflowers in the background. (Done)
The homepage of the website should be a similar format to Japan’s HQ,
with an image carousel and a menu bar grid at the bottom. (Done)
The image carousel would start with images of Suza followed by our
local images. (Done, need more local images and potentially captions)
We are allowed to keep the drop down menu from the header.